+++ py-polars/tests/unit/datatypes/test_temporal.py
@@ -817,111 +817,6 @@ def test_asof_join_tolerance_grouper() -> None:
     assert_frame_equal(out, expected)
 
 
-def test_datetime_duration_offset() -> None:
-    df = pl.DataFrame(
-        {
-            "datetime": [
-                datetime(1999, 1, 1, 7),
-                datetime(2022, 1, 2, 14),
-                datetime(3000, 12, 31, 21),
-            ],
-            "add": [1, 2, -1],
-        }
-    )
-    out = df.select(
-        [
-            (pl.col("datetime") + pl.duration(weeks="add")).alias("add_weeks"),
-            (pl.col("datetime") + pl.duration(days="add")).alias("add_days"),
-            (pl.col("datetime") + pl.duration(hours="add")).alias("add_hours"),
-            (pl.col("datetime") + pl.duration(seconds="add")).alias("add_seconds"),
-            (pl.col("datetime") + pl.duration(microseconds=pl.col("add") * 1000)).alias(
-                "add_usecs"
-            ),
-        ]
-    )
-    expected = pl.DataFrame(
-        {
-            "add_weeks": [
-                datetime(1999, 1, 8, 7),
-                datetime(2022, 1, 16, 14),
-                datetime(3000, 12, 24, 21),
-            ],
-            "add_days": [
-                datetime(1999, 1, 2, 7),
-                datetime(2022, 1, 4, 14),
-                datetime(3000, 12, 30, 21),
-            ],
-            "add_hours": [
-                datetime(1999, 1, 1, hour=8),
-                datetime(2022, 1, 2, hour=16),
-                datetime(3000, 12, 31, hour=20),
-            ],
-            "add_seconds": [
-                datetime(1999, 1, 1, 7, second=1),
-                datetime(2022, 1, 2, 14, second=2),
-                datetime(3000, 12, 31, 20, 59, 59),
-            ],
-            "add_usecs": [
-                datetime(1999, 1, 1, 7, microsecond=1000),
-                datetime(2022, 1, 2, 14, microsecond=2000),
-                datetime(3000, 12, 31, 20, 59, 59, 999000),
-            ],
-        }
-    )
-    assert_frame_equal(out, expected)
-
-
-def test_date_duration_offset() -> None:
-    df = pl.DataFrame(
-        {
-            "date": [date(10, 1, 1), date(2000, 7, 5), date(9990, 12, 31)],
-            "offset": [365, 7, -31],
-        }
-    )
-    out = df.select(
-        [
-            (pl.col("date") + pl.duration(days="offset")).alias("add_days"),
-            (pl.col("date") - pl.duration(days="offset")).alias("sub_days"),
-            (pl.col("date") + pl.duration(weeks="offset")).alias("add_weeks"),
-            (pl.col("date") - pl.duration(weeks="offset")).alias("sub_weeks"),
-        ]
-    )
-    assert out.to_dict(False) == {
-        "add_days": [date(11, 1, 1), date(2000, 7, 12), date(9990, 11, 30)],
-        "sub_days": [date(9, 1, 1), date(2000, 6, 28), date(9991, 1, 31)],
-        "add_weeks": [date(16, 12, 30), date(2000, 8, 23), date(9990, 5, 28)],
-        "sub_weeks": [date(3, 1, 3), date(2000, 5, 17), date(9991, 8, 5)],
-    }
-
-
-def test_add_duration_3786() -> None:
-    df = pl.DataFrame(
-        {
-            "datetime": [datetime(2022, 1, 1), datetime(2022, 1, 2)],
-            "add": [1, 2],
-        }
-    )
-    assert df.slice(0, 1).with_columns(
-        [
-            (pl.col("datetime") + pl.duration(weeks="add")).alias("add_weeks"),
-            (pl.col("datetime") + pl.duration(days="add")).alias("add_days"),
-            (pl.col("datetime") + pl.duration(seconds="add")).alias("add_seconds"),
-            (pl.col("datetime") + pl.duration(milliseconds="add")).alias(
-                "add_milliseconds"
-            ),
-            (pl.col("datetime") + pl.duration(hours="add")).alias("add_hours"),
-        ]
-    ).to_dict(False) == {
-        "datetime": [datetime(2022, 1, 1, 0, 0)],
-        "add": [1],
-        "add_weeks": [datetime(2022, 1, 8, 0, 0)],
-        "add_days": [datetime(2022, 1, 2, 0, 0)],
-        "add_seconds": [datetime(2022, 1, 1, 0, 0, 1)],
-        "add_milliseconds": [datetime(2022, 1, 1, 0, 0, 0, 1000)],
-        "add_hours": [datetime(2022, 1, 1, 1, 0)],
-    }
-
-
 def test_rolling_group_by_by_argument() -> None:
     df = pl.DataFrame({"times": range(10), "groups": [1] * 4 + [2] * 6})
 
@@ -2644,16 +2539,12 @@ def test_datetime_cum_agg_schema() -> None:
     assert (
         df.lazy()
         .with_columns(
-            [
-                (pl.col("timestamp").cummin()).alias("cummin"),
-                (pl.col("timestamp").cummax()).alias("cummax"),
-            ]
+            (pl.col("timestamp").cummin()).alias("cummin"),
+            (pl.col("timestamp").cummax()).alias("cummax"),
         )
         .with_columns(
-            [
-                (pl.col("cummin") + pl.duration(hours=24)).alias("cummin+24"),
-                (pl.col("cummax") + pl.duration(hours=24)).alias("cummax+24"),
-            ]
+            (pl.col("cummin") + pl.duration(hours=24)).alias("cummin+24"),
+            (pl.col("cummax") + pl.duration(hours=24)).alias("cummax+24"),
         )
         .collect()
     ).to_dict(False) == {
