+++ py-polars/tests/unit/operations/test_aggregations.py
@@ -182,29 +182,32 @@ def test_duration_function_literal() -> None:
     df = pl.DataFrame(
         {
             "A": ["x", "x", "y", "y", "y"],
-            "T": [date(2022, m, 1) for m in range(1, 6)],
+            "T": pl.datetime_range(
+                date(2022, 1, 1), date(2022, 5, 1), interval="1mo", eager=True
+            ),
             "S": [1, 2, 4, 8, 16],
         }
-    ).with_columns(
-        [
-            pl.col("T").cast(pl.Datetime),
-        ]
+    )
+
+    result = df.group_by("A", maintain_order=True).agg(
+        (pl.col("T").max() + pl.duration(seconds=1)) - pl.col("T")
     )
 
     # this checks if the `pl.duration` is flagged as AggState::Literal
-    assert df.group_by("A", maintain_order=True).agg(
-        [((pl.col("T").max() + pl.duration(seconds=1)) - pl.col("T"))]
-    ).to_dict(False) == {
-        "A": ["x", "y"],
-        "T": [
-            [timedelta(days=31, seconds=1), timedelta(seconds=1)],
-            [
-                timedelta(days=61, seconds=1),
-                timedelta(days=30, seconds=1),
-                timedelta(seconds=1),
+    expected = pl.DataFrame(
+        {
+            "A": ["x", "y"],
+            "T": [
+                [timedelta(days=31, seconds=1), timedelta(seconds=1)],
+                [
+                    timedelta(days=61, seconds=1),
+                    timedelta(days=30, seconds=1),
+                    timedelta(seconds=1),
+                ],
             ],
-        ],
-    }
+        }
+    )
+    assert_frame_equal(result, expected)
 
 
 def test_string_par_materialize_8207() -> None:
